backend/requirements.txt
Flask
qrcode[pil]
pillow
gunicorn
validators
pythonping
flask-cors